Pagination remaining

+++ src/meals/meals.service.ts
@@ -1,10 +1,12 @@
-import {Injectable, NotFoundException} from '@nestjs/common';
+import {Injectable, NotFoundException, UnauthorizedException} from '@nestjs/common';
 import {InjectRepository} from "@nestjs/typeorm";
 import {MealsRepository} from "./meals.repository";
 import {GetmealsfilterDto} from "./dto/getmealsfilter.dto";
 import {User} from "../users/user.entity";
 import {Meal} from "./meals.entity";
 import {CreatemealsDto} from "./dto/createmeals.dto";
+import {UserRole} from "../users/dto/authCredentials.dto";
+import {EditmealsDto} from "./dto/editmeals.dto";
 
 @Injectable()
 export class MealsService {
@@ -28,7 +30,15 @@ export class MealsService {
     async createMeal(createMealDto: CreatemealsDto, user: User): Promise<Meal> {
         return this.mealRepository.createMeal(createMealDto, user);
     }
-    async deleteMealById(id: number): Promise<void> {
+
+    async editMeal(editMealDto: EditmealsDto, user: User, id: number) {
+        return this.mealRepository.editMeal(id, editMealDto, user);
+    }
+    async deleteMealById(id: number, user: User): Promise<void> {
+        const found = await this.mealRepository.findOne({id});
+        if (found.userId !== user.id && user.role === UserRole.REGULAR) {
+            throw new UnauthorizedException('Regular users cannot access other users meals')
+        }
         const result = await this.mealRepository.delete({id});
         if(result.affected === 0) {
             throw new NotFoundException('Meal not found');
